Rethought of alert system
The alert type system was big and somewhat redundant, so I made it simpler.

Instead of having an enum (which used to be called Alert.Type, now Alert.Callback) that contains all different possible events to trigger when the target is met (all highlighting formats), the enum Alert.Callback contains only one HIGHLIGHT element (not 21). To precise the type of highlighting, the Callback is bundled up with a ChatFormatting : the Alert.Element class.

This explanation isn't very good, I'll explain this when we talk again ^^